Add validation requests table with polymoprhic association to our existing validation request models - This is a first step into refactoring the validation requests architecture as a whole - Create validation request records adding foreign key and type of existing data
