+++ app/models/other_change_validation_request.rb
@@ -3,7 +3,7 @@
 class OtherChangeValidationRequest < ApplicationRecord
   class ResetFeeInvalidationError < StandardError; end
 
-  include ValidationRequest
+  include ValidationRequestable
 
   belongs_to :planning_application
   belongs_to :user
