+++ spec/models/other_change_validation_request_spec.rb
@@ -199,7 +199,7 @@ RSpec.describe OtherChangeValidationRequest, type: :model do
           it "prevents an other_change_validation_request from being created" do
             expect do
               other_change_validation_request
-            end.to raise_error(ValidationRequest::ValidationRequestNotCreatableError,
+            end.to raise_error(ValidationRequestable::ValidationRequestNotCreatableError,
                                "Cannot create Other Change Validation Request when planning application has been validated")
           end
         end
